requirements.txt
# requirements.txt

requests==2.32.3
selenium==4.23.1
python-dotenv==1.0.1
scrapegraphai[burr]
nest-asyncio==1.6.0
tenacity==8.5.0
pandas==2.2.2
openpyxl==3.1.5